refactor: extract CapturedStderr helper to TestSupport
